Import `sub_0800BEF0` from kleod
Port from kleod with the struct Unk_0800BEF0 packed param (start/end XY
+ step magnitude) and gUnk_030034DC as the rolling fractional error.
Matches on first try.

- Adjust gUnk_030034DC typing to u16 (kleod-canonical; halfword load/store
  in the target asm).
- clang-format incidentals on globals.h / engine.c carried along by
  make format pre-commit.

+++ include/structs/variables.h
@@ -301,8 +301,8 @@ struct Unk_0818B8E0 {
 };
 extern struct Unk_0818B8E0 *gUnk_030051DC;
 
-/* Text-scroll Bresenham accumulator. */
-extern s32 gUnk_030034DC;
+/* Text-scroll Bresenham accumulator (mirrors kleod's u16 typing). */
+extern u16 gUnk_030034DC;
 
 /* sub_0800BEF0 (UpdateTextScroll) parameter struct. */
 struct Unk_0800BEF0 {
